# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ProЗапас API",
    "version": "1.0.0",
    "description": "HTTP-интерфейс между десктопным приложением и сервером.\n\nСистема обслуживает несколько складов одной компании. Каждый пользователь привязан к одному складу, и `warehouse_id` берётся из токена — в запросах он не передаётся. Сервер сам подставляет фильтр видимости, поэтому чужие заказы недостижимы.\n\nЗаказ — перемещение товара между двумя складами. Направление не хранится: для склада-заказчика заказ исходящий, для отправителя входящий.\n\nПодробности и обоснования — docs/api.md."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/bootstrap": {
      "get": {
        "tags": [
          "Служебные"
        ],
        "summary": "Справочники одним запросом",
        "responses": {
          "200": {
            "description": "Успех",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bootstrap"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "description": "Склады, роли, разделы, права, текущий пользователь. Один поход вместо шести при старте приложения."
      }
    },
    "/dashboard": {
      "get": {
        "tags": [
          "Служебные"
        ],
        "summary": "Сводка для главной",
        "responses": {
          "200": {
            "description": "Успех",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dashboard"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "description": "Счётчики считаются по своему складу."
      }
    },
...
    "/receipts": {
      "get": {
        "tags": [
          "Приёмка"
        ],
        "summary": "Список приёмок",
        "responses": {
          "200": {
            "description": "Успех",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "items",
                    "total",
                    "limit",
                    "offset"
                  ],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ReceiptListItem"
                      }
                    },
                    "total": {
                      "type": "integer",
                      "example": 128
                    },
                    "limit": {
                      "type": "integer",
                      "example": 50
                    },
                    "offset": {
                      "type": "integer",
                      "example": 0
                    }
                  },
                  "description": "Страница приёмок"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Unprocessable"
          }
        },
        "description": "Заказы, где мы получатель и статус shipped.",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "waiting",
                "progress",
                "done"
              ]
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Префикс номера заказа"
          },
          {
            "name": "weight_min",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "weight_max",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "created_from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "created_to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "shipped_from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "shipped_to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "maximum": 200
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0
            }
          }
        ]
      }
    },
    "/receipts/{order_id}": {
      "get": {
        "tags": [
          "Приёмка"
        ],
        "summary": "Карточка приёмки",
        "responses": {
          "200": {
            "description": "Успех",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReceiptDetail"
                }
              }
            },
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                },
                "description": "Версия документа для If-Match"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Unprocessable"
          }
        },
        "description": "Отдаёт настоящие коробки отправителя — те самые, что он упаковал. Ничего не синтезируется. Расхождение считается, не хранится.",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ]
      }
    },
    "/receipts/{order_id}/boxes/{barcode}/receive": {
      "post": {
        "tags": [
          "Приёмка"
        ],
        "summary": "Отметить коробку принятой",
        "responses": {
          "200": {
            "description": "Успех",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReceivedBox"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Unprocessable"
          }
        },
        "description": "Ключ пути — штрихкод, а не внутренний id: сканер даёт именно его. Идемпотентен: коробка уже принята — возвращаем как есть.",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "barcode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReceiveBody"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Приёмка"
        ],
        "summary": "Отменить скан",
        "responses": {
          "204": {
            "description": "Готово"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Unprocessable"
          }
        },
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "barcode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/receipts/{order_id}/complete": {
      "post": {
        "tags": [
          "Приёмка"
        ],
        "summary": "Завершить приёмку",
        "responses": {
          "200": {
            "description": "Успех",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReceiptResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Unprocessable"
          },
          "428": {
            "$ref": "#/components/responses/PreconditionRequired"
          }
        },
        "description": "Приходует принятые коробки. Коробки без отметки — зафиксированная недостача, заказ всё равно переходит в received.",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Версия документа из ETag. Без заголовка — 428, с устаревшим — 409."
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access-токен. Внутри — склад пользователя и права."
      }
    },
    "schemas": {
      "Problem": {
        "type": "object",
        "description": "RFC 9457",
        "required": [
          "type",
          "title",
          "status"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri",
            "example": "https://prozapas/errors/insufficient-stock"
          },
          "title": {
            "type": "string",
            "example": "Недостаточно свободного остатка"
          },
          "status": {
            "type": "integer",
            "example": 409
          },
          "detail": {
            "type": "string"
          }
        },
        "additionalProperties": true
      },
      "WarehouseBrief": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "code": {
            "type": "string",
            "example": "128"
          },
          "name": {
            "type": "string",
            "example": "Склад №1"
          },
          "owner": {
            "type": "string",
            "nullable": true,
            "example": "ООО «Сталкер Групп»"
          },
          "responsible": {
            "$ref": "#/components/schemas/UserBrief",
            "description": "Ответственный за склад. Карточка заказа показывает его для обеих сторон"
          }
        }
      },
      "UserBrief": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string",
            "example": "Кузнецов И.А."
          }
        }
      },
      "Person": {
        "type": "object",
        "description": "Сотрудник для фильтра: только имя и склад, без прав и почты",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string",
            "example": "Кузнецов Игорь Александрович"
          },
          "warehouse_id": {
            "type": "integer"
          }
        }
      },
...
      "Permission": {
        "type": "object",
        "required": [
          "role",
          "section",
          "can_view",
          "can_edit"
        ],
        "properties": {
          "role": {
            "type": "string",
            "example": "manager"
          },
          "section": {
            "type": "string",
            "enum": [
              "orders",
              "shipping",
              "receiving",
              "catalog",
              "stock",
              "users"
            ]
          },
          "can_view": {
            "type": "boolean"
          },
          "can_edit": {
            "type": "boolean"
          }
        }
      },
...
      "Bootstrap": {
        "type": "object",
        "properties": {
          "warehouses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WarehouseBrief"
            }
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "string"
                },
                "label": {
                  "type": "string"
                }
              }
            }
          },
          "sections": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "orders",
                "shipping",
                "receiving",
                "catalog",
                "stock",
                "users"
              ]
            }
          },
          "permissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Permission"
            }
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "warehouse": {
            "$ref": "#/components/schemas/WarehouseBrief"
          },
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Person"
            },
            "description": "Сотрудники для выпадающих фильтров. Здесь, а не в /users: тот доступен только администратору, а фильтр «Ответственный» нужен всем"
          },
          "catalog_version": {
            "type": "string"
          }
        }
      },
      "Dashboard": {
        "type": "object",
        "properties": {
          "in_work": {
            "type": "integer",
            "description": "Заказы в работе"
          },
          "outgoing_in_work": {
            "type": "integer"
          },
          "incoming_in_work": {
            "type": "integer"
          },
          "to_ship": {
            "type": "integer",
            "description": "Входящие, ждут сборки"
          },
          "to_receive": {
            "type": "integer",
            "description": "Исходящие в пути"
          },
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StatusEvent"
            }
          }
        }
      },
      "StatusEvent": {
        "type": "object",
        "properties": {
          "order_id": {
            "type": "integer"
          },
          "number": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "created",
              "processing",
              "shipped",
              "received",
              "declined",
              "cancelled"
            ]
          },
          "reason": {
            "type": "string",
            "nullable": true
          },
          "user": {
            "$ref": "#/components/schemas/UserBrief"
          },
          "occurred_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ReceiptListItem": {
        "type": "object",
        "properties": {
          "order_id": {
            "type": "integer"
          },
          "number": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "waiting",
              "progress",
              "done"
            ]
          },
          "from_warehouse": {
            "$ref": "#/components/schemas/WarehouseBrief"
          },
          "boxes_total": {
            "type": "integer"
          },
          "boxes_received": {
            "type": "integer"
          },
          "weight_expected": {
            "type": "number"
          },
          "weight_actual": {
            "type": "number"
          },
          "responsible": {
            "$ref": "#/components/schemas/UserBrief"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "shipped_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ReceivedBox": {
        "type": "object",
        "properties": {
          "barcode": {
            "type": "string"
          },
          "article": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "qty": {
            "type": "number",
            "format": "double",
            "description": "Дробное: в справочнике есть метры и листы"
          },
          "weight": {
            "type": "number",
            "description": "Вес при отгрузке"
          },
          "received_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "actual_weight": {
            "type": "number",
            "nullable": true
          },
          "diff_kg": {
            "type": "number",
            "nullable": true,
            "description": "Считается, не хранится"
          },
          "diff_percent": {
            "type": "number",
            "nullable": true
          }
        }
      },
      "ReceiveBody": {
        "type": "object",
        "required": [
          "actual_weight"
        ],
        "properties": {
          "actual_weight": {
            "type": "number",
            "example": 19.8
          }
        }
      },
      "ReceiptDetail": {
        "type": "object",
        "properties": {
          "order": {
            "$ref": "#/components/schemas/Order"
          },
          "status": {
            "type": "string",
            "enum": [
              "waiting",
              "progress",
              "done"
            ]
          },
          "responsible": {
            "$ref": "#/components/schemas/UserBrief"
          },
          "boxes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReceivedBox"
            },
            "description": "Настоящие коробки отправителя"
          },
          "version": {
            "type": "integer"
          }
        }
      },
      "ReceiptResult": {
        "type": "object",
        "properties": {
          "order": {
            "$ref": "#/components/schemas/Order"
          },
          "missing": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReceivedBox"
            },
            "description": "Коробки без отметки о приёмке"
          }
        }
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "full_name": {
            "type": "string"
          },
          "login": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "role": {
            "type": "string",
            "example": "stockman"
          },
          "warehouse": {
            "$ref": "#/components/schemas/WarehouseBrief"
          },
          "position": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "blocked"
            ]
          },
          "hire_date": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "last_login_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Тело или параметры не разобрались",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Нет токена, истёк или отозван",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Нет прав на раздел или запрещено правилом",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "NotFound": {
        "description": "Объекта нет, удалён или принадлежит чужим складам",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Conflict": {
        "description": "Конфликт версий, недопустимый переход, нехватка остатка",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Unprocessable": {
        "description": "Нарушено доменное правило",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "PreconditionRequired": {
        "description": "Изменение без If-Match",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
    }
  }
}